feat: separate Compressor and Decompressor (#11)
As part of implementing the vortex `FSSTArray`, we should separate out
the data for compression (symbol table, hash table, codes_twobyte
vector) and decompression (just the symbol table).

This PR separates the previous `SymbolTable` type into two types, a
`Compressor` and a `Decompressor<'a>`. A Compressor can be trained on a
sample text, as before, but it does not have decompression methods.
Those are now on a new `Decompressor` type. Decompressors can either be
built directly from a Compressor, or you can build one by directly
wrapping a `&[Symbol]`. Compressors allow slice access to the symbol
table as well, e.g. for serialization